Use Mash to handle string and symbol keys
Users may pass string instead of symbol in the permissions hashes, the
resource should either fail or handle them properly.

This patch propose to handle them properly using Chef Mash.
Before this patch the resource was not idempotent as during the 2nd run,
a diff was computed, something similar to:
  -'user'
  +:user